pyproject.toml
[tool.poetry]
name = "my-eat-right"
version = "0.1.0"
description = ""
authors = ["Chibuikem Iroaganachi"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
pydantic = {extras = ["email"], version = "^2.7.0"}
pydantic-extra-types = "^2.6.0"
phonenumbers = "^8.13.35"
pydantic-settings = "^2.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
alembic = "^1.13.1"
emails = "^0.6"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
